The column-width preference now makes sense in both orientations.

+++ TerpSrc/FizmoGlkDelegate.h
@@ -11,13 +11,13 @@
 #define FONTSCALE_MAX (7)
 
 @interface FizmoGlkDelegate : NSObject <IosGlkLibDelegate> {
-	CGFloat maxwidth;
+	int maxwidth; // 0 for full-width, 1 for 3/4-ish, 2 for 1/2-ish.
 	NSString *fontfamily; // as the user knows it -- not necessarily the true family name
 	int fontscale; // a number from 1 to FONTSCALE_MAX
 	int colorscheme; // 0:Bright, 1:Quiet, 2:Dark
 }
 
-@property (nonatomic) CGFloat maxwidth;
+@property (nonatomic) int maxwidth;
 /* The following properties are read from multiple threads, so we declare them atomic. */
 @property (retain) NSString *fontfamily;
 @property int fontscale;
